src/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22
python-dotenv==1.2.2
pydantic-settings==2.13.1

langchain==1.2.12
langchain-community==0.4.1
langchain-huggingface==1.2.1
langchain-groq==1.1.2
langchain-cohere==0.5.0
langchain-google-genai
langchain-core

sentence-transformers==5.4.0
google-generativeai

qdrant-client[fastembed]
flashrank==0.2.10

ragas==0.4.3

aiofiles==25.1.0
httpx

pypdf==6.9.1

redis==4.5.5
python-telegram-bot
tavily-python == 0.7.23
langgraph == 1.1.6

langgraph-checkpoint-sqlite ==  3.0.3
